Marauder's Map Milestone 1
This is the first part of a 3 part overview for Marauder's Map - a Human Computer Interaction group project. Although this document was originally written in part by other group members, I have revised and rewritten everything other than the section in italics. Additionally, I have used some of the methodologies and templates from the Interaction Design Foundation to improve the format of the personas.